requirements.txt
# AI Finance Advisor - Complete Application Dependencies

# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and AI
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
google-generativeai>=0.3.0

# Vector Database
faiss-cpu==1.7.4
sentence-transformers>=2.2.2

# Document Processing
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.3

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Security and Validation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS and Middleware
fastapi-cors==0.0.6

# Logging and Monitoring
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# Production
gunicorn==21.2.0
